Refactor: Move old services into a new folder to prepare new services update
