+++ src/app/(Main)/TeachingUnit/page.tsx
@@ -10,7 +10,7 @@ import {
 	SelectValue,
 } from "@/components/ui/select";
 import { initialLevels } from "@/lib/niveau_utils";
-import { getMatterForLevel } from "@/server/Matter";
+import { getMatterForLevel } from "@/server_old/Matter";
 import { FileText } from "lucide-react";
 
 import { useEffect, useState } from "react";
