+++ src/server_old/data/Teacher.json
@@ -26,7 +26,6 @@
 	{ "name": "Patrice", "Abr_Teacher": "PAT" },
 	{ "name": "RALAIVAO Jean-Christian", "Abr_Teacher": "RAL" },
 	{ "name": "Roger", "Abr_Teacher": "ROG" },
-	{ "name": "Siaka", "Abr_Teacher": "SIA" },
 	{ "name": "Venot", "Abr_Teacher": "VEN" },
 	{ "name": "Victor", "Abr_Teacher": "VIC" },
 	{ "name": "Volatiana", "Abr_Teacher": "VOL" },
